# src/main/resources/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: FoodMind Backend API
  version: v1
  description: Public API contract for FoodMind clients.
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    CorrelationId:
      description: Accepted or generated request correlation identifier.
      schema:
        type: string
        maxLength: 128
      example: postman-correlation-test
  parameters:
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      description: Optional client-generated correlation identifier.
      schema:
        type: string
        maxLength: 128
      example: postman-correlation-test
...
    Page:
      name: page
      in: query
      required: false
      description: Zero-based page index.
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      required: false
      description: Bounded page size.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - email
        - displayName
        - password
        - clientType
      properties:
        email:
          type: string
          format: email
          maxLength: 320
          example: primary@example.test
        displayName:
          type: string
          maxLength: 100
          example: Primary User
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
          example: correct horse battery
        timeZone:
          type: string
          maxLength: 64
          example: Asia/Singapore
        clientType:
          type: string
          enum:
            - WEB
            - ANDROID
          example: WEB
        deviceLabel:
          type: string
          maxLength: 100
          example: Pixel 9
    LoginRequest:
      type: object
      required:
        - email
        - password
        - clientType
      properties:
        email:
          type: string
          format: email
          maxLength: 320
          example: primary@example.test
        password:
          type: string
          format: password
          maxLength: 128
          example: correct horse battery
        clientType:
          type: string
          enum:
            - WEB
            - ANDROID
          example: WEB
        deviceLabel:
          type: string
          maxLength: 100
          example: Chrome on Windows
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
          maxLength: 4096
          description: Opaque refresh token for Android/Postman clients. Web clients may use the Secure HttpOnly cookie instead.
          example: redacted-refresh-token
        clientType:
          type: string
          enum:
            - WEB
            - ANDROID
          example: ANDROID
    AuthTokenResponse:
      type: object
      required:
        - userId
        - accessToken
        - tokenType
        - expiresIn
        - expiresAt
        - refreshToken
        - refreshTokenExpiresAt
        - csrfToken
      properties:
        userId:
          type: string
          format: uuid
          example: 018fd1e4-5e5d-7a30-90d4-8d0d3f7f5d9c
        accessToken:
          type: string
          description: Short-lived signed JWT access token.
          example: redacted-access-jwt
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          format: int64
          example: 900
        expiresAt:
          type: string
          format: date-time
          example: "2026-07-29T12:30:00Z"
        refreshToken:
          type: string
          description: Opaque refresh token returned once; only a SHA-256 hash is stored by the backend.
          example: redacted-refresh-token
        refreshTokenExpiresAt:
          type: string
          format: date-time
          example: "2026-08-28T12:15:00Z"
        csrfToken:
          type: string
          description: CSRF companion value for cookie-based web refresh/logout.
          example: redacted-csrf-token
...
    HistoryResponse:
      type: object
      required:
        - fromUtcInclusive
        - toUtcExclusive
        - period
        - types
        - timeZone
        - buckets
        - entries
      properties:
        fromUtcInclusive:
          type: string
          format: date-time
          description: UTC lower bound derived once from the requested local or offset boundary.
        toUtcExclusive:
          type: string
          format: date-time
          description: UTC exclusive upper bound derived once from the requested local or offset boundary.
        period:
          type: string
          enum:
            - DAY
            - WEEK
            - MONTH
        types:
          type: array
          items:
            type: string
            enum:
              - FOOD
              - DRINK
        timeZone:
          type: string
          example: Asia/Singapore
        buckets:
          type: array
          items:
            $ref: '#/components/schemas/HistoryBucket'
        entries:
          type: array
          items:
            $ref: '#/components/schemas/HistoryEntry'
        nextCursor:
          type: string
          nullable: true
          description: Opaque cursor using occurredAt DESC, sourceType ASC, sourceId DESC ordering.
...
    CookingIngredientRequest:
      type: object
      required:
        - ingredientName
      properties:
        ingredientName:
          type: string
          minLength: 1
          maxLength: 160
          example: Firm tofu
        quantity:
          type: number
          exclusiveMinimum: 0
          nullable: true
          example: 300
        unit:
          type: string
          maxLength: 40
          nullable: true
          example: g
        source:
          type: string
          enum:
            - MANUAL
            - AUTHORISED_PANTRY
          default: MANUAL
    GenerateCookingPlanRequest:
      type: object
      required:
        - servings
      properties:
        ingredients:
          type: array
          maxItems: 30
          items:
            $ref: '#/components/schemas/CookingIngredientRequest'
        recipeIds:
          type: array
          maxItems: 25
          items:
            type: string
            format: uuid
        servings:
          type: integer
          minimum: 1
          maximum: 24
          example: 2
        maxMinutes:
          type: integer
          minimum: 1
          maximum: 1440
          nullable: true
          example: 60
        maxBudget:
          type: number
          minimum: 0
          nullable: true
          example: 20
        currency:
          type: string
          minLength: 3
          maxLength: 3
          nullable: true
          example: SGD
        requiredDietaryTagCodes:
          type: array
          items:
            type: string
          example:
            - VEGAN
        avoidAllergenCodes:
          type: array
          items:
            type: string
          example:
            - PEANUT
        servingAt:
          type: string
          format: date-time
          nullable: true
          example: '2026-08-02T18:30:00+08:00'
        region:
          type: string
          maxLength: 8
          nullable: true
          example: SG
    CookingPlanResponse:
      type: object
      required:
        - planId
        - status
      properties:
        planId:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - PROCESSING
            - READY
            - NEEDS_CONFIRMATION
            - INFEASIBLE
            - FAILED
        planRevision:
          type: string
          nullable: true
          example: 3f4a…:v1
        region:
          type: string
          nullable: true
          example: SG
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        solverStatus:
          type: string
          nullable: true
          example: OPTIMAL
        makespanMinutes:
          type: integer
          nullable: true
          example: 54
        errorCode:
          type: string
          nullable: true
          example: SCHEDULE_UNKNOWN
        errorMessage:
          type: string
          nullable: true
        sources:
          type: array
          items:
            $ref: '#/components/schemas/CookingPlanSource'
        timeline:
          type: array
          items:
            $ref: '#/components/schemas/CookingPlanTimelineTask'
        miseEnPlace:
          type: array
          items:
            $ref: '#/components/schemas/CookingPlanMiseEnPlaceItem'
        dishCompletions:
          type: array
          items:
            $ref: '#/components/schemas/CookingPlanDishCompletion'
        completionChecklist:
          type: array
          items:
            $ref: '#/components/schemas/CookingPlanCompletionItem'
        assumptions:
          type: array
          items:
            $ref: '#/components/schemas/CookingPlanAssumption'
        repairOptions:
          type: array
          items:
            $ref: '#/components/schemas/CookingPlanRepairOption'
        questions:
          type: array
          items:
            type: string
        confirmationQuestions:
          type: array
          items:
            $ref: '#/components/schemas/CookingPlanConfirmationQuestion'
        decisions:
          type: array
          items:
            $ref: '#/components/schemas/CookingPlanDecision'
        reasons:
          type: array
          items:
            type: string
        safeAlternatives:
          type: array
          items:
            type: string
        safetyPolicy:
          allOf:
            - $ref: '#/components/schemas/CookingPlanSafetyPolicy'
          nullable: true
        explanation:
          type: string
          nullable: true
        explanationSource:
          type: string
          nullable: true
          example: deterministic
...
    CookingPlanDishCompletion:
      type: object
      properties:
        dishId:
          type: string
        completionMinute:
          type: integer
          minimum: 0
        taskCount:
          type: integer
          minimum: 0
        isShared:
          type: boolean
...
    CookingQuestionAnswer:
      type: object
      required:
        - questionId
        - value
      properties:
        questionId:
          type: string
          minLength: 1
          maxLength: 128
        value:
          type: string
          minLength: 1
          maxLength: 500
...
    CookingPlanHistoryResponse:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/CookingPlanSummary'
...
  responses:
    BadRequest:
      description: Request validation or JSON parsing failed.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Authentication is required or invalid.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
    NotFound:
      description: Resource was not found or cannot be disclosed.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Conflict:
      description: Request conflicts with current resource state.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
paths:
...
  /cooking-plans/generate:
    post:
      tags:
        - Cooking Plans
      operationId: generateCookingPlan
      summary: Generate a controlled cooking plan.
      description: Requires Idempotency-Key. The backend persists the request, resolves hard dietary/allergy rules, sends only active controlled recipe candidates to the Cooking Agent outside a database transaction, validates the structured result, and stores an immutable output snapshot.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateCookingPlanRequest'
      responses:
        '201':
          description: Cooking plan created or resolved from a matching idempotent retry.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CookingPlanResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /cooking-plans/{planId}:
    get:
      tags:
        - Cooking Plans
      operationId: getCookingPlan
      summary: Read an owned cooking plan.
      description: Owner-only read returning immutable request and output snapshots.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - name: planId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Persisted cooking plan.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CookingPlanResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /cooking-plans/{planId}/decisions:
    post:
      tags:
        - Cooking Plans
      operationId: submitCookingPlanDecisions
      summary: Submit decisions for a NEEDS_CONFIRMATION plan.
      description: Validates the answers against the presented confirmation questions, guards the plan revision (409 when the plan is not awaiting confirmation), and re-runs the agent to produce a new revision plan.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - name: planId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CookingQuestionAnswer'
      responses:
        '200':
          description: The new revision plan produced from the submitted decisions.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CookingPlanResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /cooking-plans/history:
    get:
      tags:
        - Cooking Plans
      operationId: listCookingPlanHistory
      summary: List owned cooking plans.
      description: Returns only the authenticated user's plans using bounded pagination and stable createdAt/id ordering.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Owned cooking plan page.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CookingPlanHistoryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a user and issue tokens.
      security: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered and session created.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
            Set-Cookie:
              description: Secure HttpOnly refresh cookie and CSRF companion cookie for web clients.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate with email and password.
      security: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Credentials accepted and session created.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
            Set-Cookie:
              description: Secure HttpOnly refresh cookie and CSRF companion cookie for web clients.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /history:
    get:
      tags:
        - History
      summary: Query combined timezone-aware food and drink history.
      description: Uses [from, to) range semantics. Date-only boundaries are interpreted in the resolved IANA timezone; week buckets start on Monday. Ranges over 366 days are rejected.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - name: from
          in: query
          required: true
          schema:
            type: string
          example: "2026-07-26"
        - name: to
          in: query
          required: true
          schema:
            type: string
          example: "2026-07-28"
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum:
              - DAY
              - WEEK
              - MONTH
            default: DAY
        - name: types
          in: query
          required: false
          description: Comma-separated source types.
          schema:
            type: string
            example: FOOD,DRINK
        - name: timeZone
          in: query
          required: false
          schema:
            type: string
            example: Asia/Singapore
        - name: groupId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: cuisineId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: placeId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Presentation-neutral combined history.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Rotate a refresh session and issue a new access token.
      security: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Refresh token rotated.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
            Set-Cookie:
              description: Replacement refresh and CSRF cookies for web clients.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Revoke the supplied refresh session.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '204':
          description: Refresh session revoked and auth cookies cleared.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
            Set-Cookie:
              description: Expired refresh and CSRF cookies.
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout-all:
    post:
      tags:
        - Authentication
      summary: Revoke every active refresh session for the authenticated user.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '204':
          description: Active refresh sessions revoked and auth cookies cleared.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
            Set-Cookie:
              description: Expired refresh and CSRF cookies.
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
...